Reuse structured logging for API events
